Port "rename without explicitly specifying database"
